pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.13",
    "jose>=1.0.0",
    "passlib>=1.7.4",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "supabase>=2.16.0",
    "uvicorn>=0.34.3",
]
